Internal Aalen's Additive Hazards Model storing a linear effect estimate for each event time

Do not call this function on its own

Usage

Areg(out.formula, id, data, method, ...)

Arguments

out.formula

Survival formula for Aalen's additive hazards model.

id

character string indicating which column of 'data' corresponds to the subject ID.

data

Data set in counting process format. In particular the data should contain a "start", "stop" and "event" column along with any mediators and baseline covariates.

method

passed from dpa call, defaults to "timereg", otherwise "aareg"

...

other parameters passed to Aalen's additive hazards regression function "timereg::aalen()"

Value

data.frame with observation times and estimated coefficients for independent variables in "regformula"

Examples

library(dpasurv)

data(simdata)

obj <- Areg(Surv(start,stop,event)~M+x, id="subject", data=simdata, method="timereg")