Internal function that gathers meta data on the input to function dpa

Do not call this function on its own

Usage

get.meta(out.formula, mediator.formulas, data)

Arguments

out.formula

obtained directly from corresponding input to dpasurv::dpa

mediator.formulas

obtained directly from corresponding input to dpasurv::dpa

data

obtained directly from corresponding input to dpasurv::dpa

Value

this function returns meta data associated with the call to dpasurv::dpa